Improved interaction with program
Program now set up better. It runs and asks for a command, and takes
input relevant to the command when a command is entered.

Now more commands can be worked in easily.

Next command to work in is modified. Takes a directory and date and
returns the name of all directories/files within that directory that
have been modified after the specified date